124_001 R1604_0049 Error en fase de pruebas. Para probar que se podía introducir un long en una consulta y en su informe asociado, he creado una consulta y he cambiado por BD el idconsulta a 4380000000. Este número es superior al máximo de integer. Al acceder a la pestaña de Configuracion da error. Es más, al intentar crear otra consulta a continuación, da error y he visto en el log que precisamente da error al intentar convertir a Integer.
Por otra parte, considero que esta solución es una chapuza, porque el número actual de consultas por colegio no llega ni a las 1000 (y ya sería demasiado). Así que lo q yo haría sería precisamente lo contrario: limitar las consultas a 4 dígitos, y es más que suficiente.
Mostrar menos

+++ src/com/siga/beans/AdmConsultaInformeAdm.java
@@ -69,7 +69,7 @@ public class AdmConsultaInformeAdm extends MasterBeanAdministrador {
 			bean.setIdInstitucion(UtilidadesHash.getInteger(hash, AdmConsultaInformeConsultaBean.C_IDINSTITUCION));
 			bean.setIdPlantilla(UtilidadesHash.getString(hash, AdmConsultaInformeConsultaBean.C_IDPLANTILLA));
 			bean.setIdInstitucion_consulta(UtilidadesHash.getInteger(hash, AdmConsultaInformeConsultaBean.C_IDINSTITUCION_CONSULTA));
-			bean.setIdConsulta(UtilidadesHash.getInteger(hash, AdmConsultaInformeConsultaBean.C_IDCONSULTA));
+			bean.setIdConsulta(UtilidadesHash.getLong(hash, AdmConsultaInformeConsultaBean.C_IDCONSULTA));
 			bean.setNombre(UtilidadesHash.getString(hash, AdmConsultaInformeConsultaBean.C_NOMBRE));
 			bean.setVariasLineas(UtilidadesHash.getString(hash, AdmConsultaInformeConsultaBean.C_VARIASLINEAS));
 			bean.setDescripcion(UtilidadesHash.getString(hash, AdmConsultaInformeConsultaBean.C_DESCRIPCION));
